backend/requirements.txt
Flask
Flask-CORS
Pillow
google-genai
python-dotenv
gunicorn
requests